agents/requirements.txt
langchain==0.3.18
langchain_community==0.3.17
langchain-openai==0.3.4
openai==1.61.1
streamlit==1.42.0
python-dotenv==1.0.1
wikipedia==1.4.0
wikipedia-api==0.8.1
google-search-results==2.4.2